Create the Code
Import urllib, BeautifulSoup, and ssl, then request user to input url or use a default url if the user leaves the prompt blank (<1 character in length). Open and read the data and run through BeautifulSoup. The code then retrieves all of the span tags, looping through the data and adding the span tags to a list. Using list comprehension, it converts the list of lists to a list of strings. Then, loops through the list of strings, converting the strings to integers and counts each item and adds the number to the sum. Finally, print the final count and sum.